按需调整 int 为 long

+++ src/main/java/net/wuxianjie/springrestapi/role/Role.java
@@ -7,13 +7,13 @@ import java.time.LocalDateTime;
 @Data
 public class Role {
 
-  private Integer id;
+  private Long id;
   private LocalDateTime createdAt;
   private LocalDateTime updatedAt;
   private String remark;
   private String name;
   private String menus; // 英文逗号分隔的权限字符串
-  private Integer parentId; // 上级角色 id
+  private Long parentId; // 上级角色 id
   private String parentName; // 上级角色名
   private String fullPath; // 节点全路径, 以英文句号分隔的角色 id 字符串
 }
